Adding the unit tests for the kMCPDFT
